Fix climber configuration parsing

+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/sketch/climbing/SimpleClimber.java
@@ -100,7 +100,7 @@ final class SimpleClimber implements HillClimber {
       return config().getDouble(BASE_PATH + "percent-sample");
     }
     public double tolerance() {
-      return config().getInt(BASE_PATH + "tolerance");
+      return config().getDouble(BASE_PATH + "tolerance");
     }
   }
 }
